added equality comparison operators, changed offset_type from uint16_t to uint32_t in case of overflows

+++ include/flatmemory/details/layout_utils.hpp
@@ -31,7 +31,7 @@ namespace flatmemory
 /**
  * Common data types
  */
-using offset_type = uint16_t;
+using offset_type = uint32_t;
 
 using buffer_size_type = uint32_t;
